+++ allauth/socialaccount/views.py
@@ -3,8 +3,11 @@ from django.contrib.auth.decorators import login_required
 from django.contrib.sites.shortcuts import get_current_site
 from django.http import HttpResponseRedirect
 from django.urls import reverse, reverse_lazy
+from django.utils.decorators import method_decorator
 from django.views.generic.base import TemplateView
 from django.views.generic.edit import FormView
+
+from allauth.account.decorators import reauthentication_required
 
 from ..account import app_settings as account_settings
 from ..account.adapter import get_adapter as get_account_adapter
@@ -93,6 +96,13 @@ class LoginErrorView(TemplateView):
 login_error = LoginErrorView.as_view()
 
 
+@method_decorator(
+    reauthentication_required(
+        allow_get=True,
+        enabled=lambda request: account_settings.REAUTHENTICATION_REQUIRED,
+    ),
+    name="dispatch",
+)
 class ConnectionsView(AjaxCapableProcessFormViewMixin, FormView):
     template_name = "socialaccount/connections." + account_settings.TEMPLATE_EXTENSION
     form_class = DisconnectForm